Move app/networks into networks folder
