+++ webapp/context/evmWalletContext.tsx
@@ -8,8 +8,8 @@ import {
 } from '@rainbow-me/rainbowkit'
 import { metaMaskWallet } from '@rainbow-me/rainbowkit/wallets'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { allEvmNetworks } from 'app/networks'
 import { useNetworks } from 'hooks/useNetworks'
+import { allEvmNetworks } from 'networks'
 import { useMemo } from 'react'
 import { http } from 'viem'
 import { WagmiProvider, createConfig } from 'wagmi'
